Address PR comments - Move classes, methods
